Clean-up: remove experimental code; fix imports; constants and comments
