--- rust-src/concordium_base/src/curve_arithmetic/ed25519_arkworks.rs
@@ -1,7 +1,0 @@
-use super::arkworks_instances::CurveElementLength;
-use ark_curve25519::*;
-
-impl CurveElementLength for EdwardsProjective {
-    const GROUP_ELEMENT_LENGTH: usize = 64;
-    const SCALAR_LENGTH: usize = 32;
-}
